Add the three F5529 Assembly clocking examples
F5529Clocks1: Flash the LED 1.0 with a delay of 250000 cycles to the
default 1MHz clock. MCLK and SMCLK brought out to pins
F5529Clocks2: Connect XT1 and XT2 to F5529. XT1 is used as a source for
ACLK and XT2 for MCLK and SMCLK. (both brought out to pins). LED 1.0
flash with 250000 cycles delay
F5529Clocks3: Use DCO/FLL to clock F5529 to 21MHz. Flash LED 1.0 with 1
million cycles delay